### 2024-05-05 #6723 @editorialbot
[PRE REVIEW]: SVInsight: A Python Package for Calculating Social Vulnerability Indices
**Submitting author:** @mdp0023 (<a href="http://orcid.org/0000-0002-5149-3744">Matthew Preisser</a>)
**Repository:** https://github.com/mdp0023/SVInsight
**Branch with paper.md** (empty if default branch): main
**Version:** v1.1.0
