### 2024-05-24 #6723 @mdp0023
Thank you for the query scope. For reference, the commit history (first commit date and number of commits) won’t be reflective of the packages development history as it was previously only stored locally and not in a public GitHub repository. Prior to the creation of a GitHub repo for this code base, it was used in academic publications going back to 2021 including multiple in progress manuscripts.
