### 2024-05-24 #6723 @crvernon
Thanks for the extra info @mdp0023!  In that case, I am going to pull it from scope query.  The editor that is best suited to take on this submission is currently at full capacity, so I am going to set this to waitlist for now and get you an editor assigned ASAP.
