### 2024-07-01 15bd2e9
Update functionality to include 2022 data
