+++ svinsight/__init__.py
@@ -1,3 +1,3 @@
-__version__= "1.1.0"
+__version__= "1.1.1"
 
 from svinsight.svi import SVInsight
