### 2024-07-25 #6723 @mdp0023
If it's possible we would like to update the version to be reviewed to v1.1.1, we had a minor patch applied a few weeks back.
