### 2024-08-01 #6723 @mstimberg
@samhforbes I am afraid I have to decline at this time – I am about to leave for summer holidays and will only be fully back to office in September…
