+++ svinsight/__init__.py
@@ -1,3 +1,3 @@
-__version__= "1.1.1"
+__version__= "1.1.2"
 
 from svinsight.svi import SVInsight
