+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "poetry.core.masonry.api"
 
 [tool.poetry]
 name = "svinsight"
-version = "1.1.1"
+version = "1.1.2"
 description = "Create social vulnerability index"
 authors = ["Matthew Preisser <mattpreisser@gmail.com>"]
 license = "MIT"
