### 2024-08-23 #6723 @sappelhoff
@samhforbes thanks for reaching out, I'll be happy to edit this.

Hi @mdp0023 👋 do you have suggestions as to whom I shall invite as reviewers? (if in line with our [COI policies](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy))
