### 2024-08-23 #6723 @sappelhoff
Thanks for the suggestions @mdp0023!

👋 @jamesdamillington @FATelarico @mashrur-ayon would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
