### 2024-09-08 #6723 @sappelhoff
👋 @dataspider @LCB0B @bkrayfield would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
