### 2024-09-08 #6723 @dataspider
@sappelhoff I'm happy to review this but cannot guarantee full execution until the second half of October; let me know if this works for you. Meanwhile, I took a quick pass through the companion paper and would encourage the authors (@mdp0023) to go through the review guidelines you linked above themselves and ensure that they provide enough evidence on the individual checklist items in their write-up to go through the review process smoothly.
