### 2024-10-19 mdp0023/SVInsight#4 @dataspider
Author contributions 
From the review checklist:
> Does the full list of paper authors seem appropriate and complete?

The paper lists all three authors with equal contributions, but only the first author has a significant commit history.
Could you clarify the authors' contributions (which can of course also be equal without equal commit history)?
I'd recommend a contribution statement at the end of the manuscript, ideally following the [CREDiT Taxonomy](https://credit.niso.org/).

While you are at it, please also add the university name(s) to your affiliations.

https://github.com/openjournals/joss-reviews/issues/7212
