### 2024-10-19 mdp0023/SVInsight#5 @dataspider
JOSS Text Edits
- Global: "user defined" -> "user-defined"
- l. 7: "A community's exposure" -> "The exposure of communities" (the plural also makes. the "their" in that sentence correct; otherwise, you'd need to change that to "its")
- l. 14: comma before "including"
- l. 17: "region specific" -> "region-specific" + comma before "including" + "including" -> "such as" (to avoid repetition)
- l. 25: "vulnerability indices are not created equally" -> "not all vulnerability indices are created equal"
- ll. 32–34: This sentence should go into a separate section "Projects Using SVInsight" (and it could be expanded a bit).
- l. 43: Cut-off sentence: "socio-demographic va Furthermore"
- l. 44: "taken" -> "interpreted"
- ll. 46/47: word repetition "utilize" (perhaps use "leverage" as a replacement for the second occurrence)
- ll. 67/68: "the locations with the higher ... have the highest ..." -> "locations with higher ...  have higher ..."
- l. 70: comma after "Texas"
- ll. 70/71: "distinction ... **is** visible"
- ll. 71–73: "used in various research settings" -> add citations
- l. 73: "Tate et al. (2021)" -> "(Tate et. al., 2021)"
- l. 85/86: "fills missing holes as needed" – please clarify what this means in the text
- ll. 88/89: This seems unnecessarily coupled (it would be better to have functions to compute different SVIs separately -> I will elaborate on this during my code review)
- ll. 90–92: It would be good to have functionality for adding your own index computation methodology (I will elaborate on this during my code review)
- Before l. 93: any related packages that could be mentioned in a "Related Packages" section?
- Before l. 93: insert contribution statement (see #4)
- l. 147: ensure caps in "United States"

https://github.com/openjournals/joss-reviews/issues/7212
