### 2024-10-24 d952b0f
Add contributing statement
