### 2024-10-24 8abb94b
add university affiliations
