### 2024-10-24 mdp0023/SVInsight#4 @mdp0023
Thank you for the feedback. I added a contribution statement, `d952b0f`, and the university affiliations, `8abb94b`
