+++ joss/paper.md
@@ -39,20 +39,20 @@ bibliography: paper.bib
 
 # Summary
 
-A community's exposure to environmental hazards, their sensitivity to such events, and their ability to respond (adaptive capacity) are influenced by their social, political, cultural, economic, and demographic information [@Smit_2006; @Cutter_2010; @Fatemi_2017]. Understanding the interconnected relationships among exposure, sensitivity, and adaptive capacity is important to estimate the degree to which stakeholders can mitigate environmental hazards [@Smit_2006]. Social Vulnerability Indices, or SVIs, are built on social and demographic data to serve as proxies for these interconnected variables. Numerous national and international SVIs exist including SoVI &reg; from The University of South Carolina's Hazards Vulnerability & Resilience Institute [@Cutter_2003], the Center for Disease Control [@Flanagan_2011], and the United Nations Development Program [@UNDP_2010], as well as more region specific indicators including Texas A&M's Hazard Reduction and Recovery Center's Coastal Indicator [@Peacock_2010] and the Austin Area Sustainability Indicator [@Bixler_2021a]. In this paper, we present an open-source Python package, *SVInsight*, which provides an accessible workflow to calculate social vulnerability indices that are specific to user defined study areas while using the two leading calculation methodologies. 
+The exposure of communities to environmental hazards, their sensitivity to such events, and their ability to respond (adaptive capacity) are influenced by their social, political, cultural, economic, and demographic information [@Smit_2006; @Cutter_2010; @Fatemi_2017]. Understanding the interconnected relationships among exposure, sensitivity, and adaptive capacity is important to estimate the degree to which stakeholders can mitigate environmental hazards [@Smit_2006]. Social Vulnerability Indices, or SVIs, are built on social and demographic data to serve as proxies for these interconnected variables. Numerous national and international SVIs exist including, SoVI &reg; from The University of South Carolina's Hazards Vulnerability & Resilience Institute [@Cutter_2003], the Center for Disease Control [@Flanagan_2011], and the United Nations Development Program [@UNDP_2010], as well as more region-specific indicators, such as Texas A&M's Hazard Reduction and Recovery Center's Coastal Indicator [@Peacock_2010] and the Austin Area Sustainability Indicator [@Bixler_2021a]. In this paper, we present an open-source Python package, *SVInsight*, which provides an accessible workflow to calculate social vulnerability indices that are specific to user-defined study areas while using the two leading calculation methodologies. 
 
 
 # Statement of need
-Generic social vulnerability indices for large regions can be lacking in their ability to identify at risk populations [@Tate_2012; @Nelson_2015; @Tellman_2020]. Furthermore, vulnerability indices are not created equally, and users, researchers, and developers should clearly state the objectives and structure of their index in order to accurately present their findings [@Bakkensen_2016]. *SVInsight* is an accessible and open-source tool to quickly calculate SVIs for a user defined region using either custom or a research supported pre-set list of social and demographic variables. *SVInsight* calculates SVIs built on data from the Census Bureau's 5-Year American Community Survey using the two leading methods for calculating such metrics: a composite score using a data reduction methodology [@Cutter_2003] and a ranking method [@Flanagan_2011]. The work done leading up to the creation of the *SVInsight* Python package was and continues to be a part of numerous vulnerability and hazard related research efforts [@Bixler_2021a; @Preisser_2022; @Preisser_2023].
[…]
