### 2024-11-02 mdp0023/SVInsight#5 @dataspider
@mdp0023 The changes don't seem to be in the latest editorialbot-generated PDF, please double-check that they are reflected in the latest version at https://github.com/openjournals/joss-reviews/issues/7212
