### 2024-11-02 mdp0023/SVInsight#7 @dataspider
Contribution Guide URL in README
There's a typo in your README that causes the contribution guide to be "not found" (note the extra "l" after "html"):

> We welcome contributions to SVInsight. Please open an issue or a pull request if there is functionality you would like to see or propose. Refer to our [contributing guide](https://mdp0023.github.io/SVInsight/Contributions/contributions.htmll) for more information.

https://github.com/openjournals/joss-reviews/issues/7212
