### 2024-11-02 mdp0023/SVInsight#6 @dataspider
Also, please add pytest to the dev dependencies in your pyproject.toml and be at least one level more specific than "*" for pinning dependencies.
