### 2024-11-02 mdp0023/SVInsight#8 @dataspider
Running Tests
It seems that to run _any_ of the tests, I need an API key for the census data – although there is quite a bit of functionality that could be tested simply with the test data you deliver with the package.
Please separate the tests into those that handle data-retrieval functionality (which will require an API key) and tests that handle computational functionality, given locally available data.

https://github.com/openjournals/joss-reviews/issues/7212
