### 2024-11-02 mdp0023/SVInsight#9 @dataspider
OS- and API-key dependencies in examples
Please make the examples independent of your operating system and your API key.

You have lines like `file_path = "/Users/matthewpreisser/Documents/Research/Codes/SVInsight"` (looks like MacOS) in `svinsight_example_Harris_County.py` or `file_path = "C:/Users/pp7545/Documents/research/SVInsight"` (looks like Windows) in `SVInsight_example_SETx.ipynb` (you might also want to include jupyterlab in the dev dependencies, btw).
Instead of those, programmatically generate the absolute paths to match the OS your code is run on or tell people to create some kind of .env file to specify your root and read the env via a suitable library.

Regarding the API key: You are redistributing the census data for two countries with the repository, so it seems that you could easily use these data to demonstrate your functionality without making API calls (and then just have an API call demo in a separate example). I don't want to have to get an API key just to try out some software (I know that many people working in your field will have one anyway, but relying on an API key in the examples is unnecessary, so why not make the package more accessible?).
Please also check the data licensing situation for the data that you are redistributing, and include the applicable licensing terms in a DATALICENSE file or as a separate section of your license file.

https://github.com/openjournals/joss-reviews/issues/7212
