### 2024-11-05 mdp0023/SVInsight#10 @bkrayfield
Substantial scholarly effort
This is not meant to be a slight to all of the great work done on this repository, but can the authors explain more about why this is different from just a database?

To elaborate, the package seems to be computing the social vulnerability index always from the Census API. So, why not just create a database where this is already provided for all the data in the Census API? Is there a reason I would expect the data to change? So if I pull the data for New York at a specific time, will it change from user to user?

https://github.com/openjournals/joss-reviews/issues/7212
