### 2024-11-05 mdp0023/SVInsight#10 @bkrayfield
To add to this comment, would it make sense to add more options to allow for custom borders, etc?
