### 2024-11-05 mdp0023/SVInsight#11 @bkrayfield
Cited Lit
The current paper has 17 cited papers, but the docs only show 7 papers. Is this an error?

https://github.com/openjournals/joss-reviews/issues/7212
