### 2024-11-05 mdp0023/SVInsight#8 @bkrayfield
I agree with this, a small testing file would be useful.
