### 2024-11-05 mdp0023/SVInsight#12 @bkrayfield
State of the field
Are there any other options out there? I am not sure, but this is a requirement of the paper. With the authors of the software package not being the primary "method" creators, I would like to see the authors include some additional text about how people have been doing this research before this package.

https://github.com/openjournals/joss-reviews/issues/7212
