+++ joss/paper.bib
@@ -199,7 +199,7 @@
 @Article{Tate_2021,
   author    = {Eric Tate and Md Asif Rahman and Christopher T. Emrich and Christopher C. Sampson},
   journal   = {Natural Hazards},
-  title     = {Flood exposure and social vulnerability in the {United States},
+  title     = {Flood exposure and social vulnerability in the {United States}},
   year      = {2021},
   month     = {jan},
   number    = {1},
