### 2024-11-08 728c75e
Update dependency versions
