+++ pyproject.toml
@@ -12,15 +12,17 @@ readme = "long_desc.md"
 
 [tool.poetry.dependencies]
 python = "^3.9"
-census = "*"
-factor_analyzer = "*"
-geopandas = "*"
-numpy = "*"
-pandas = "*"
-PyYAML = "*"
-scikit_learn = "*"
-openpyxl = "*"
-matplotlib = "*"
+census = ">=0.8.22"
+factor_analyzer = ">=0.5.1"
+geopandas = ">=1.0.1"
+numpy = ">=2.1"
+pandas = ">=2.2.3"
+PyYAML = ">=6.0.2"
+scikit_learn = ">=1.5.2"
+openpyxl = ">=3.1.5"
+matplotlib = ">=3.9.2"
+setuptools = ">=75.3"
 
 [tool.poetry.dev-dependencies]
 # Add your development dependencies here
+pytest = ">=8.3.3"
