### 2024-11-08 dacaf5b
Add gdal installation disclaimer
