### 2024-11-09 23f5985
reconfigure tests
