+++ tests/test_svinsight_functionality.py
@@ -1,0 +1,219 @@
+# Testing functionality of svinsight without requiring API key
+
+# import package
+from svinsight import SVInsight as svi 
+import os
+import pytest
+import geopandas as gpd
+import matplotlib.figure
+import sys
+
+
+
+# Get the directory that contains this file
+this_directory = os.path.dirname(os.path.abspath(__file__))
+
+# one level up 
+dir1 = os.path.dirname(this_directory)
+# two levels up 
+dir2 = os.path.dirname(os.path.dirname(this_directory))
+
+# Add the root directory to sys.path
+sys.path.insert(0, this_directory)
+sys.path.insert(0, dir1)
+sys.path.insert(0, dir2)
+
+
+
+
+##############################################################################################
+# TESTS FOR MINOR PROJECT FUNCTIONALITY
+
+# set some initial value with blank api_key
+project_name='test_proj'
+file_path=os.getcwd()
+geoids=['48453']
+api_key = ''
+
+
+def test_svi():
+    """Test that project can be created"""
+    assert svi(f"{os.path.dirname(os.path.realpath(__file__))}/{project_name}", file_path, api_key, geoids)
+
+def test_svi_project_name():
+    """Test that project name is string"""
+    with pytest.raises(TypeError):
+        svi(2, file_path, api_key, geoids)
+
+def test_svi_file_path_exists():
+    """Test that file path exists"""
+    with pytest.raises(FileNotFoundError):
+        svi(project_name, 'incorrect_file_path', api_key, geoids)
+
+def test_svi_geoids_incorrect():
+    """Test various scenarios of incorrect geoids"""
+    with pytest.raises(ValueError):
+        geos=[['123'],
+              ['a','b','c'],
+              [12345,12],
+              ['12345','48']]
+        for geo in geos:
+            svi(project_name, file_path, api_key, geo)
+
+def test_svi_geoids_correct():
+    """Test various scenarios of correct geoids"""
+    geos=[['12', '48'],
+            ['12345','48256']]
+    for geo in geos:
+       assert svi(project_name, file_path, api_key, geo)
+
+
+
+
+
+
+
+##############################################################################################
+# TESTS FOR MAJOR PROJECT FUNCTIONALITY
+# create a generic test architecture
+def run_svi_workflow(project_name, file_path, geoids, config, boundary, year, overwrite=False, varchange=False):
+    """Run the svi workflow for a given project"""
+    project = svi(project_name, file_path, api_key, geoids)
+
+    if varchange is True:
+        # change variables to calculate an economic index
+         # check configure
+        try:
+            project.configure_variables(f"{config}_{year}_{boundary}", include=['MDHSEVAL','PERCAP','QRICH'])
+        except Exception:
+            assert False, "configure_variables execution failed"
+
+    else:
+        # check configure
+        try:
+            project.configure_variables(f"{config}_{year}_{boundary}")
+        except Exception:
+            assert False, "configure_variables execution failed"
+
+    # check if output files exist
+    output_file3 = os.path.join(project.variables, f"{config}_{year}_{boundary}.yaml")
+    assert os.path.isfile(output_file3)
+    
+    # check svi creation
+    try:
+        project.calculate_svi(config_file = f"{config}_{year}_{boundary}", boundary=boundary, year=year)
+    except:
+        assert False, "calculate_svi execution failed"
+
+    # Check that output files exist 
+    output_file4 = os.path.join(project.svis, f"{project.project_name}_{year}_{boundary}_{config}_{year}_{boundary}_svi.gpkg")
+    output_file5 = os.path.join(project.svis, f"{project.project_name}_{year}_{boundary}_{config}_{year}_{boundary}_svi.csv")
[…]
