+++ tests/test_svinsight.py
@@ -1,202 +1,0 @@
-
-import sys
-import os
-
-# Get the directory that contains this file
-this_directory = os.path.dirname(os.path.abspath(__file__))
-
-# one level up 
-dir1 = os.path.dirname(this_directory)
-# two levels up 
-dir2 = os.path.dirname(os.path.dirname(this_directory))
-
-# Add the root directory to sys.path
-sys.path.insert(0, this_directory)
-sys.path.insert(0, dir1)
-sys.path.insert(0, dir2)
-
-# import package
-from svinsight import SVInsight as svi 
-import os
-import pytest
-import geopandas as gpd
-
-# set some initial value
-project_name='test_proj'
-file_path=os.getcwd()
-api_key=os.environ.get('API_KEY')
-geoids=['48453']
-
-def test_svi():
-    """Test that project can be created"""
-    assert svi(project_name, file_path, api_key, geoids)
-
-def test_svi_project_name():
-    """Test that project name is string"""
-    with pytest.raises(TypeError):
-        svi(2, file_path, api_key, geoids)
-
-def test_svi_file_path_exists():
-    """Test that file path exists"""
-    with pytest.raises(FileNotFoundError):
-        svi(project_name, 'incorrect_file_path', api_key, geoids)
-
-def test_svi_geoids_incorrect():
-    """Test various scenarios of incorrect geoids"""
-    with pytest.raises(ValueError):
-        geos=[['123'],
-              ['a','b','c'],
-              [12345,12],
-              ['12345','48']]
-        for geo in geos:
-            svi(project_name, file_path, api_key, geo)
-
-def test_svi_geoids_correct():
-    """Test various scenarios of correct geoids"""
-    geos=[['12', '48'],
-            ['12345','48256']]
-    for geo in geos:
-       assert svi(project_name, file_path, api_key, geo)
-
-
-
-# create a generic test architecture
-def run_svi_workflow(project, boundary, year, overwrite=True, varchange=False):
-    # test project instance
-    assert project
-    # test folder instances
-    for loc in ['Boundaries','Variables', 'Data', 'Documentation', 'SVIs']:
-        assert os.path.exists(os.path.join(project.file_path, loc))
-    
-    # Call the function with the test inputs
-    gpkg = project.boundaries_data(boundary=boundary, year=year, overwrite=overwrite)
-
-    # Check that the result is a GeoDataFrame
-    assert isinstance(gpkg, gpd.GeoDataFrame)
-
-    # Check that the output file exists
-    output_file = os.path.join(project.boundaries, f"{project.project_name}_{year}_{boundary}.gpkg")
-    assert os.path.isfile(output_file)
-
-    # check extraction of raw census data (if successful, returns nothing)
-    try:
-        project.census_data(boundary, year, interpolate=True, verbose=True, overwrite=overwrite)
-    except Exception:
-        assert False, "census_data execution failed"
-
-    # check that the output files exists
-    output_file1 = os.path.join(project.data, f"{project.project_name}_{year}_{boundary}_rawdata.csv")
-    output_file2 = os.path.join(project.data, f"{project.project_name}_{year}_{boundary}_rawdata.gpkg")
-    assert os.path.isfile(output_file1)
-    assert os.path.isfile(output_file2)
-
-    if varchange is True:
-        # change variables to calculate an economic index
-         # check configure
-        try:
-            project.configure_variables(f"{config}_{year}_{boundary}", include=['MDHSEVAL','PERCAP','QRICH'])
-        except Exception:
-            assert False, "configure_variables execution failed"
-
-    else:
-        # check configure
-        try:
-            project.configure_variables(f"{config}_{year}_{boundary}")
-        except Exception:
[…]
