### 2024-11-09 fbfc679
add jupyterlab as dev dependency
