+++ pyproject.toml
@@ -26,3 +26,4 @@ setuptools = ">=75.3"
 [tool.poetry.dev-dependencies]
 # Add your development dependencies here
 pytest = ">=8.3.3"
+jupyterlab = ">=3.0.0"
