### 2024-11-09 f06225a
updates to examples and data license
