+++ examples/svinsight_example_Travis_County.py
@@ -1,6 +1,12 @@
 import sys
 import os
 
+# This package comes with initial Census boundaries and data for Travis County, TX
+# With overwrite=False, the package will not download the data again 
+# Therefore, an empty api_key will work for this example
+# However, to download the data, you will need to get an API key from the US Census Bureau
+
+
 # Get the directory that contains this file
 this_directory = os.path.dirname(os.path.abspath(__file__))
 
@@ -21,8 +27,9 @@ import matplotlib.pyplot as plt
 
 # set variables
 project_name = 'Travis_County'
-file_path = "/Users/matthewpreisser/Documents/Research/Codes/SVInsight"
-api_key = os.environ.get('API_KEY')
+file_path = dir1
+# api_key = os.environ.get('API_KEY')
+api_key=''
 geoids = ['48453']
 
 # create instance
