+++ joss/paper.bib
@@ -219,3 +219,36 @@
   groups    = {SVI},
   publisher = {Texas A&M University},
 }
+
+@article{VanZandt_2012,
+  title={Mapping social vulnerability to enhance housing and neighborhood resilience},
+  author={Van Zandt, Shannon and Peacock, Walter Gillis and Henry, Dustin W and Grover, Himanshu and Highfield, Wesley E and Brody, Samuel D},
+  journal={Housing Policy Debate},
+  volume={22},
+  number={1},
+  pages={29--55},
+  year={2012},
+  publisher={Taylor \& Francis}
+}
+
+@article{Xu_2024,
+  title={findSVI: an R package to calculate the Social Vulnerability Index at multiple geographical levels},
+  author={Xu, Heli and Li, Ran and Bilal, Usama},
+  journal={Journal of open source software},
+  volume={9},
+  number={99},
+  pages={6525},
+  year={2024}
+}
+
+
+@article{Cutter_2008,
+  title={Temporal and spatial changes in social vulnerability to natural hazards},
+  author={Cutter, Susan L and Finch, Christina},
+  journal={Proceedings of the national academy of sciences},
+  volume={105},
+  number={7},
+  pages={2301--2306},
+  year={2008},
+  publisher={National Acad Sciences}
+}
