### 2024-11-10 79fa9b5
fix python dependency for PyPI
