+++ svinsight/__init__.py
@@ -1,3 +1,3 @@
-__version__= "1.1.3"
+__version__= "1.1.4"
 
 from svinsight.svi import SVInsight
