### 2024-11-10 001fad4
allow for manual publishing to PyPI
