+++ .github/workflows/publish_pypi.yml
@@ -5,6 +5,8 @@ on:
  push: 
    tags:
      - "v*.*.*"
+ workflow_dispatch:
+    
 jobs:
   build:
     runs-on: ubuntu-latest
