+++ joss/paper.md
@@ -43,7 +43,7 @@ The exposure of communities to environmental hazards, their sensitivity to such 
 
 
 # Statement of need
-Generic social vulnerability indices for large regions can be lacking in their ability to identify at risk populations [@Tate_2012; @Nelson_2015; @Tellman_2020]. Furthermore, not all vulnerability indices are created equal, and users, researchers, and developers should clearly state the objectives and structure of their index in order to accurately present their findings [@Bakkensen_2016]. Research utilizing SVIs has historically been limited to either existing national or regional databases [@Cutter_2003; @Flanagan_2011; VanZandt_2012; Bixler_2021a]. Researchers that want to incorporate social vulnerability information into their research are therefore limited to these existing databases and their pre-determined variable choices, or developing their own index from scratch. Furthermore, researchers’ understanding of what contributes to vulnerability is variable in both time and space [@Cutter_2008]. Therefore, it is necessary to tailor SVIs to the context in which they are being applied. 
+Generic social vulnerability indices for large regions can be lacking in their ability to identify at risk populations [@Tate_2012; @Nelson_2015; @Tellman_2020]. Furthermore, not all vulnerability indices are created equal, and users, researchers, and developers should clearly state the objectives and structure of their index in order to accurately present their findings [@Bakkensen_2016]. Research utilizing SVIs has historically been limited to either existing national or regional databases [@Cutter_2003; @Flanagan_2011; @VanZandt_2012; @Bixler_2021a]. Researchers that want to incorporate social vulnerability information into their research are therefore limited to these existing databases and their pre-determined variable choices, or developing their own index from scratch. Furthermore, researchers’ understanding of what contributes to vulnerability is variable in both time and space [@Cutter_2008]. Therefore, it is necessary to tailor SVIs to the context in which they are being applied. 
 
 *SVInsight* is an accessible and open-source tool to quickly calculate SVIs for a user-defined region using either custom or a research supported pre-set list of social and demographic variables. This package creates a pipeline between extensively large datasets and easily customizable SVIs, allowing researchers to experiment and manipulate various indices (e.g., general vulnerability, economic, race/ethnicity, etc.) built on a variety of variables more efficiently and effectively.*SVInsight* calculates SVIs built on data from the Census Bureau's 5-Year American Community Survey using the two leading methods for calculating such metrics: a composite score using a data reduction methodology [@Cutter_2003] and a ranking method [@Flanagan_2011]. There is a recently published package in R that is capable of calculating the Rank Method SVI [@Xu_2024], but to the best of the author’s knowledge there is no open source factor analysis based SVI package.
 
