### 2024-11-10 mdp0023/SVInsight#6 @mdp0023
Thank you for the feedback. I had trouble recreating your issue, but I am aware that gdal can be a difficult dependency to install. I therefore added some information on the installation guide about this potential requirement (depending on the user’s coding environment). Please let me know if this is insufficient and I can work on additional documentation. I also added the most recent versions of the dependencies and added pytest as a development dependency.
