### 2024-11-10 mdp0023/SVInsight#8 @mdp0023
Thank you for the suggestion, I also agree that would add a lot of functionality to the tests. I split the original tests into two separate files, one to test functionality for data that comes with the package, and another that tests the data acquisition functions. I also set them up to remove intermediary files after tests are conducted to reduce unnecessary file sizes.”
