### 2024-11-10 mdp0023/SVInsight#9 @mdp0023
Thank you for the suggestion and I agree this will increase the accessibility of the package for first time users. Since the data download functions won’t redownload the input data with the default overwrite parameter (overwrite=False), and as you mentioned the data is provided with the package, the examples should work now without requiring a current API. I also removed the file paths and made them absolute paths, thank you for catching that. The appropriate language has also been added to the license with regard to the included data.
