### 2024-11-10 mdp0023/SVInsight#10 @mdp0023
Thank you for your feedback, and I have added some material to the paper manuscript to address your comment in the “statement of need” section. To summarize: you are correct, the data for a given location will likely not change whether it is pulled today or one year from now (barring any statistical changes from the US Census Bureau, but that is not common). However, what does change throughout time is our understanding of what variables contribute to vulnerability both spatially and temporally. As a result, it is necessary for researchers to tailor SVIs to the context in which they are being applied. Therefore, the purpose of this package is to create the pipeline between extensively large datasets and easily customizable SVIs, based on the two leading methods to do such. This will allow researchers to experiment and manipulate various indices (e.g., general vulnerability, economic, race/ethnicity, etc.) built on a variety of variables more efficiently and effectively. With regard to your comment about custom borders, this is also a feature of the package. Users can select whatever counties or states they are interested in, and based on this decision, the final SVI will be different, especially when using the factor analysis method which is highly dependent on input data. For example, the vulnerability variables are weighted differently based on their significance for whichever region is being examined.
