### 2024-11-10 mdp0023/SVInsight#11 @mdp0023
My apologies, could you clarify what docs only show 7 papers? I’m seeing the correct count on the paper, the DOIs on the Joss review thread, and in the paper.bib file. Happy to fix whatever is missing.
