### 2024-11-10 mdp0023/SVInsight#12 @mdp0023
Thank you for your feedback, I have provided more information within the manuscript to address this, as well as #10 , which I believe are slightly related. To summarize here: Research utilizing SVIs has historically been limited to either existing national or regional databases. Researchers that want to incorporate social vulnerability information into their research are therefore limited to these existing databases and their pre-determined variable choices, or developing their own index from scratch. Furthermore, researchers’ understanding of what contributes to vulnerability is variable in both time and space. Therefore, it is necessary to tailor SVIs to the context in which they are being applied. The purpose of this package is to create the pipeline between extensively large datasets and easily customizable SVIs, based on the two leading SVI creation methods. This will allow researchers to experiment and manipulate various indices (e.g., general vulnerability, economic, race/ethnicity, etc.) built on a variety of variables more efficiently and effectively. There is a recently published package in R that is capable of calculating the Rank Method SVI, but to the best of the author’s knowledge there is no open source factor analysis based SVI package.
