### 2024-11-16 #7212 @sappelhoff
> Thank you to the reviewers for their comments so far, I believe I have addressed them all so far in the latest version of the package (V1.2). I am happy to provide additional information or address additional comments as necessary.

Thanks for the update @mdp0023!

@bkrayfield @dataspider -- could you please update us with a tentative timeline of when you plan to finalize your reviews?
