### 2024-11-24 mdp0023/SVInsight#6 @dataspider
Thank you for the updates; I would probably still add the conda workaround I sketched above somewhere in the docs (I'd keep it simple in the README and point to a more elaborate `Installation` section in your online docs for trouble-shooting).
Also, there is an "it's/its" typo in the installation instructions now that you may want to fix :)
