### 2024-11-26 mdp0023/SVInsight#6 @mdp0023
Sounds good, I will add your more specific workaround (and other possible solutions) to the docs website and keep it simple in the README. Thanks for the typo find too
