### 2024-11-26 1f8f63c
installation guide updates
