+++ _sources/Acknowledgements/acknowledgements.rst.txt
@@ -1,0 +1,11 @@
+Acknowledgements
+================
+
+We would like to thank the many collaborators who have helped us developing our methodologies. This work has been incorporated into numerous projects that have been supported by the following grants:
+
++ National Science Foundation Graduate Research Fellowship (grant no. DGE-1610403)
++ Future Investigators in NASA Earth and Space Science and Technology (NASA FINESST, grant no. 21-EARTH21-0264)
++ Planet Texas 2050, a research grand challenge at the University of Texas at Austin.
++ U.S. Department of Energy, Office of Science, Biological and Environmental Research Program’s South-East Texas Urban Integrated Field Laboratory under Award Number DE-SC0023216.
+
+
