+++ _sources/User_Guide/userguide.rst.txt
@@ -1,0 +1,235 @@
+.. currentmodule:: SVInsight.svi
+
+User Guide
+==========
+
+Overview
+--------
+
+The basic workflow of `SVInsight` works as follows: 
+
+1. Initialize the project class (:obj:`SVInsight`) with your project name, folder path, Census API key, and your study area.
+
+2. Define the initial project configuration variables with a boundary (tract or block groups) and year.
+
+3. Extract the necessary boundary geopackage using :meth:`SVInsight.boundaries_data`.
+
+4. Extract the raw census data and fill in holes in the data using :meth:`SVInsight.census_data`.
+
+5. Configure the run and calculate the SVI using :meth:`SVInsight.configure_variables` and :meth:`SVInsight.calculate_svi`. 
+
+
+Below you will find the detailed user guide that walks through each of these steps, with additional options for project customization. `SVInsight` is supposed to be a flexible and exploratory estimator of various social vulnerability indices, and is therefore highly customizable within the framework of Factor Analysis and Rank Method methodologies. The User Guide below walks through the methods available for calculating SVIs. For a larger discussion on the Factor Analysis and Rank Method methodologies, please refer to the :doc:`Background </Background/background>` and  :doc:`Understanding Iterative Factor Analysis Results </Background/understanding>` pages. 
+
+
+1. Initializing the Project
+---------------------------
+
+The first step to the project is importing :obj:`SVInsight` and setting the project `params.` Each instance of the class can be seen as a standalone project for the input geoids, where users can configure different runs based on years, boundaries, and variables of interest. In this User Guide we are using Travis_County, Texas as a working example.
+
+.. doctest::
+
+   >>> from svinsight import SVInsight as svi 
+   >>> project_name = 'Travis_County'
+   >>> file_path = "/User/Documents"
+   >>> api_key = os.environ.get('API_KEY')
+   >>> geoids = ['48453']
+   >>> proj = svi(project_name, file_path, api_key, geoids)
+
+These variables create the instance of a project. Each have the following purpose: 
+
+- ``project_name``: This is the name of your project. In this case, it's 'Travis_County'. The project name is used when saving files throughout the project.
+- ``file_path``: This is the path to the parent folder where your project files will be stored. Creating the SVInsight object automatically creates the necessary file structure needed for your project based on the ``file_path`` and ``project_name``. 
+- ``api_key``: This is your Census API key. You can obtain an API key from the `Census Bureau's developer page <https://www.census.gov/data/developers/data-sets.html>`_.
+- ``geoids``: This is a list of geographic identifiers for the areas you're interested in. In this case, it's ['48453'], which represents Travis County, Texas. The necessary geographic identifiers for your state or county of interest can be obtained from this `FCC page <https://transition.fcc.gov/oet/info/maps/census/fips/fips.txt>`_. 
+
+During the initialization, a series of folder is generated for storing the files created during a project. The folder locations and their intended contents are as follows:
+
+- **Boundaries:** Will contain geopackages of the boundary areas based on the input geoids.
+- **Data:** Where the raw census data csv and geopackages will be stored, in addition to other csv files that contain information about missing variables and how they were rectified. 
+- **Documentation:** Will contain excel sheets of information pertinent to the factor analysis svi. 
+- **SVIs:** Where the geopackages of SVIs will be stored.
+- **Variables:** Where the configuration files will be stored.
+
+
+2. Defining the Initial Project Configuration Variables
+-------------------------------------------------------
+
[…]
