+++ _sources/apiref/apiref.rst.txt
@@ -1,0 +1,7 @@
+SVInsight
+=========
+A Census API key is needed to run this package. You can obtain an API key from the `Census Bureau's developer page <https://www.census.gov/data/developers/data-sets.html>`_.
+
+.. automodule:: svinsight.svi
+    :members:
+    :special-members:
