+++ _sources/Examples/examples.rst.txt
@@ -1,0 +1,13 @@
+Examples
+========
+
+Bellow you will find a Jupyter Notebook that computes an SVI for the Southeast Texas region. More example scripts can be found on the `GitHub page in the Examples folder <https://github.com/mdp0023/SVInsight/tree/main/examples>`_.
+
+
+
+Example Notebooks
+#################
+.. toctree::
+   :maxdepth: 2
+    
+   SVInsight_example_SETx.ipynb

[…]
