### 2025-02-16 #7212 @sappelhoff
Dear @mdp0023,

sorry for the radio silence on this submission.

In the background (via email, and the JOSS editor Slack), there have been some discussions regarding the "[substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort)" criterion that we have at JOSS, and whether or not [SVInsight](https://github.com/mdp0023/SVInsight) fulfills the requirements.

These discussions were including the peer reviewers @dataspider and @bkrayfield, who brought this issue to my attention, as well as the associate editor in chief for this track, @samhforbes, who in turn also brought the discussed issues up with his colleague AEiCs.

At JOSS, we usually conduct "scope reviews" to assess whether or not a submission fulfills all requirements to proceed to peer review. In the present case, the submission was voted "in scope", however during the (deeper-diving) peer review process, several shortcomings were identified. Below, I try to summarize the most important points:

- SVInsight seems to be close to a "wrapper" around the census API with added utilities for social vulnerability scores.
- SVInsight seems to duplicate functionalities already available in other packages like the census package and factor_analyzer.
- SVInsight has issues such as unnecessary coupling of functionalities and hardcoding of data (see also https://github.com/mdp0023/SVInsight/issues/10).

Overall, the submission was judged to require significant refactoring to become a useful utility package.

For JOSS, this presents itself as a difficult tradeoff, as on the one hand, we want peer-reviewers to point to required improvements, rather than recommending a "rejection". However, on the other hand, a major refactoring can require a lot of effort which can fall outside of the scope of a reviewer to support (reviewer burden). Some related information has been discussed here: https://github.com/openjournals/joss/issues/1395

I personally agree with the assessments of the peer review process. It is unclear how we can proceed now. One option that I see is that I step down as an editor for this submission and @samhforbes finds a suitable replacement. Such a replacement would ideally be an editor with more domain knowledge for this particular submission than I have, and a willingness to go through several bigger changes as part of the review process. @samhforbes please let me know what you would prefer.

@dataspider and @bkrayfield I would be very grateful if you could communicate your intentions on whether you want to continue the peer-review here, or not. In case of the latter, we can simply unassign you from this issue.

Thank you for everyone who offered their opinion on this.
Thank you @dataspider and @bkrayfield for all the effort you have already put into the review up to this point.
And thank you for your patience @mdp0023.
