### 2025-02-17 #7212 @mdp0023
We’d like to thank the editors and the reviewers for the time they have put into this review so far, and we’d like to take this opportunity to respond to some of your comments.

We believe that this package does in fact represent a significant scholarly effort as per JOSS guidelines and we would like to make that case one more time in the context of your most recent comments.

SVInsight’s functionality does not currently exist and therefore fills a need. SVInsight is intended to:

- streamline the creation of exploratory social vulnerability indices (SVIs).
- calculate SVIs from US Census data using two research supported calculation methods. One method is our own iterative factor analysis workflow that has been in development within our research group for over two years, and the second uses a rank method that was made popular by the CDC.
- allow for comparisons between SVI frameworks during the exploratory phase of research. Given user defined socio-economic variables and geographic areas of interest, unique SVIs can be calculated within seconds or minutes.

> SVInsight seems to duplicate functionalities already available in other packages like the census package and factor_analyzer.

SVInsight utilizes the census and factor_analyzer to access Census data and conduct the actual factor analysis. We found that these performed well and there was no need to duplicate the work. SVInsight adds the functionality and methodology to create SVIs that these packages do not have.

> SVInsight seems to be close to a "wrapper" around the census API with added utilities for social vulnerability scores.

SVInsight utilizes an existing wrapper for the Census API and adds utility for calculating social vulnerability scores, which represents the significant scholarly effort. The initial creation of these methods (i.e., to calculate an SVI for a single county) and the subsequent package took well over a year to create. With SVInsight, indices for entire states are calculated in less than a few minutes. This issue was already addressed in the pre-review of the paper.

> SVInsight has issues such as unnecessary coupling of functionalities and hardcoding of data

You referenced mdp0023/SVInsight#10 and this is not hardcoding. As we responded in that thread, it is not feasible to create an SVI database of all Census areas because the SVI estimate itself is dependent on the study area and what variables are included. SVInsight gives users the flexibility to define their index (including variables and geographic extent) based on their interests. If you are referring to a different instance of hardcoding data, we would appreciate a further description of the specific issue so that we can address it.

There are two specific points on the JOSS Guidelines for what constitutes scholarly effort that we believe SVInsight falls under:

- Whether the software has already been cited in academic papers, and
- Whether the software is sufficiently useful that it is likely to be cited by other researchers working in this domain.

This package, and the methods behind it, have already been utilized in three published papers ([Bixler et al., 2021](https://doi.org/10.1016/j.ijdrr.2021.102613), [Preisser et al., 2022](https://doi.org/10.5194/hess-26-3941-2022), [Preisser et al., 2023](https://doi.org/10.3389/frwa.2023.1278205)), one under review paper, and at least two in preparation papers. The reason we sought to publish the package itself is for the intention of being able to cite the work. We additionally know that the package is being used by other research groups outside of our own. We are aware that "single-function packages are not acceptable" according to the JOSS Guidelines. While SVInsight serves a single purpose, it is not a single-function package, and it represents a streamlined package that can aid researchers in their ability to incorporate multiple iterations of custom SVIs into their own work.
