### 2025-03-31 #7212 @arfon
Firstly, I want to acknowledge the efforts of everyone involved in this review process. JOSS relies upon the dedication of volunteer reviewers and editors, and the thoughtful engagement of authors. Your contributions are all deeply appreciated.

The track Editor-in-Chief, Samuel Forbes, has asked me to weigh in on this submission and help bring the discussion to a close. I also note that the authors reached out to the editorial team via email to request intervention from the Editor-in-Chief.

While I recognize that some communications between the reviewers, editor, and authors could have been clearer and more consistent—particularly regarding expectations around scope and documentation—I believe the overall evaluations and decision reached are sound. The summary [comment from the editor](https://github.com/openjournals/joss-reviews/issues/7212#issuecomment-2661349885) accurately reflects the consensus of the reviewers and editorial team.

This submission was considered “borderline” on several fronts, including the novelty of the implementation, the completeness of documentation, and its alignment with the JOSS [scope and review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html). Based on the unanimous recommendation of the reviewers, editor, and track EiC, the decision is to reject this submission.

To be clear, not being accepted by JOSS does not mean your software is without value. It simply means that it did not fully meet our publication criteria at this time.

As an alternative, you may wish to follow [GitHub’s guide](https://guides.github.com/activities/citable-code/) to create a permanent archive and DOI for your software. This allows others to cite your work even without a formal journal publication.

Arfon Smith, Editor-in-Chief, Journal of Open Source Software
